refactor: extract ReportHeader and Gauge components
