+++ src/components/report/ReportHeader.tsx
@@ -1,0 +1,113 @@
+import React from 'react';
+import {
+    Activity,
+    Calendar,
+    User,
+    Ruler,
+    Zap
+} from 'lucide-react';
+import Gauge from './Gauge';
+
+interface ReportHeaderProps {
+    client: any;
+    measurement: any;
+    ffmi: {
+        value: number;
+        label: string;
+        description: string;
+        color: string;
+    };
+}
+
+export default function ReportHeader({ client, measurement, ffmi }: ReportHeaderProps) {
+    const getActivityLevelLabel = (level: number | null) => {
+        if (!level) return '--';
+        const labels: Record<number, string> = {
+            1: 'Sedentario',
+            2: 'Moderadamente activo',
+            3: 'Muy activo / Atleta'
+        };
+        return labels[level] || `Nivel ${level}`;
+    };
+
+    return (
+        <div className="bg-plum rounded-[3rem] p-12 text-white shadow-2xl relative overflow-hidden">
+            <div className="relative z-10 flex flex-col md:flex-row md:items-center justify-between gap-8">
+                <div className="flex-1">
+                    <h1 className="text-5xl font-bold mb-6">{client.name} {client.lastname}</h1>
+                    <div className="grid grid-cols-2 lg:grid-cols-3 gap-y-6 gap-x-8">
+                        <div className="flex items-center gap-3">
+                            <User size={20} className="text-white/50" />
+                            <div>
+                                <p className="text-[10px] font-bold uppercase tracking-widest opacity-50">Género</p>
+                                <p className="font-semibold">{client.gender === 'male' ? 'Hombre' : 'Mujer'}</p>
+                            </div>
+                        </div>
+                        <div className="flex items-center gap-3">
+                            <Calendar size={20} className="text-white/50" />
+                            <div>
+                                <p className="text-[10px] font-bold uppercase tracking-widest opacity-50">Edad</p>
+                                <p className="font-semibold">{client.age || '--'} años</p>
+                            </div>
+                        </div>
+                        {client.birthday && (
+                            <div className="flex items-center gap-3">
+                                <Calendar size={20} className="text-white/50" />
+                                <div>
+                                    <p className="text-[10px] font-bold uppercase tracking-widest opacity-50">Cumpleaños</p>
+                                    <p className="font-semibold">
+                                        {new Date(client.birthday).toLocaleDateString('es-ES', { day: 'numeric', month: 'long' })}
+                                    </p>
+                                </div>
+                            </div>
+                        )}
+                        <div className="flex items-center gap-3">
+                            <Ruler size={20} className="text-white/50" />
+                            <div>
+                                <p className="text-[10px] font-bold uppercase tracking-widest opacity-50">Altura</p>
+                                <p className="font-semibold">{measurement.height || client.height || '--'} cm</p>
+                            </div>
+                        </div>
+                        <div className="flex items-center gap-3">
+                            <Activity size={20} className="text-white/50" />
+                            <div>
+                                <p className="text-[10px] font-bold uppercase tracking-widest opacity-50">Nivel de Actividad</p>
+                                <p className="font-semibold">{getActivityLevelLabel(measurement.activityLevel || client.activityLevel)}</p>
+                            </div>
+                        </div>
+                        {client.sessionsPerWeek && (
+                            <div className="flex items-center gap-3">
+                                <Zap size={20} className="text-white/50" />
+                                <div>
+                                    <p className="text-[10px] font-bold uppercase tracking-widest opacity-50">Sesiones por semana</p>
+                                    <p className="font-semibold">{client.sessionsPerWeek} sesiones</p>
+                                </div>
+                            </div>
+                        )}
+                    </div>
+                </div>
+                <div className="bg-white/10 backdrop-blur-md rounded-3xl p-3 border border-white/20 h-fit min-w-[340px] flex flex-col items-center justify-center">
+                    <Gauge
+                        value={ffmi.value}
+                        min={12} max={30}
+                        unit="kg/m²"
+                        markers={client.gender === 'female' ? [
+                            { label: 'Bajo', val: 15 },
+                            { label: 'Promedio', val: 18 },
+                            { label: 'Excelente', val: 22 },
+                            { label: 'Superior', val: 30 }
+                        ] : [
+                            { label: 'Bajo', val: 18 },
+                            { label: 'Promedio', val: 21 },
+                            { label: 'Excelente', val: 25 },
+                            { label: 'Superior', val: 30 }
+                        ]}
+                    />
+                    <div className="text-[10px] font-bold uppercase tracking-widest opacity-70 mt-4 text-center leading-tight">
+                        Índice de Masa<br />Libre de Grasa
+                    </div>
+                </div>
+            </div>
+        </div>
+    );
+}
